+++ src/test/java/spleetwaise/address/logic/LogicManagerTest.java
@@ -21,7 +21,6 @@ import spleetwaise.address.storage.JsonAddressBookStorage;
 import spleetwaise.address.testutil.Assert;
 import spleetwaise.address.testutil.PersonBuilder;
 import spleetwaise.address.testutil.TypicalPersons;
-import spleetwaise.commons.IdUtil;
 import spleetwaise.commons.core.GuiSettings;
 import spleetwaise.commons.exceptions.SpleetWaiseCommandException;
 import spleetwaise.commons.logic.commands.CommandResult;
@@ -31,6 +30,7 @@ import spleetwaise.commons.model.CommonModel;
 import spleetwaise.commons.model.UserPrefs;
 import spleetwaise.commons.storage.JsonUserPrefsStorage;
 import spleetwaise.commons.storage.StorageManager;
+import spleetwaise.commons.util.IdUtil;
 import spleetwaise.transaction.model.ReadOnlyTransactionBook;
 import spleetwaise.transaction.model.TransactionBookModel;
 import spleetwaise.transaction.model.TransactionBookModelManager;
