+++ src/test/java/spleetwaise/transaction/storage/adapters/JsonAdaptedTransactionTest.java
@@ -14,8 +14,8 @@ import spleetwaise.address.model.AddressBookModelManager;
 import spleetwaise.address.model.person.Person;
 import spleetwaise.address.storage.JsonAdaptedPerson;
 import spleetwaise.address.testutil.TypicalPersons;
-import spleetwaise.commons.IdUtil;
 import spleetwaise.commons.exceptions.IllegalValueException;
+import spleetwaise.commons.util.IdUtil;
 import spleetwaise.transaction.model.transaction.Date;
 import spleetwaise.transaction.model.transaction.Description;
 import spleetwaise.transaction.model.transaction.Transaction;
